requirements.txt
PyQt5==5.15.10
openai==1.8.0
pygame==2.5.2
python-dotenv==1.0.0
httpx==0.26.0
